+++ app/components/GettingStartedModal/steps/SecretsStep/index.ts
@@ -1,0 +1,3 @@
+import SecretsStep from './SecretsStep';
+
+export default SecretsStep;
